refactor(mcp): split stdio bootstrap from server tools
