+++ packages/mcp-server/src/index.ts
@@ -1,797 +1,2 @@
-#!/usr/bin/env node
-
-/**
- * @codemem/mcp — MCP stdio server.
- *
- * Runs as a separate process spawned by the host (OpenCode/Claude).
- * Owns its own better-sqlite3 connection. Communicates via stdio JSON-RPC.
- *
- * Port of codemem/mcp_server.py — all 13 tools.
- */
-
-import { spawn } from "node:child_process";
-import { existsSync } from "node:fs";
-import { dirname, join } from "node:path";
-import {
-	dedupeOrderedIds,
-	type MemoryItemResponse,
-	type MemoryResult,
-	MemoryStore,
-	projectMatchesFilter,
-	resolveDbPath,
-	storeVectors,
-	VERSION,
-} from "@codemem/core";
-import { McpServer } from "@modelcontextprotocol/sdk/server/mcp.js";
-import { StdioServerTransport } from "@modelcontextprotocol/sdk/server/stdio.js";
-import { z } from "zod";
-import {
-	forgetMemoryForMcp,
-	getManyForMcp,
-	getMemoryForMcp,
-	rememberMemoryForMcp,
-} from "./memory-access.js";
-import { buildFilters, resolveDefaultProject } from "./project-scope.js";
-
-// ---------------------------------------------------------------------------
-// Static data
-// ---------------------------------------------------------------------------
-
-const MEMORY_KINDS: Record<string, string> = {
-	discovery: "Something learned about the codebase, architecture, or tools",
-	change: "A code change that was made",
-	feature: "A new feature that was implemented",
-	bugfix: "A bug that was found and fixed",
-	refactor: "Code that was refactored or restructured",
-	decision: "A design or architecture decision",
-	exploration: "An experiment or investigation (may not have shipped)",
-};
-
-// ---------------------------------------------------------------------------
-// Shared zod filter schema (spread into each tool that accepts filters)
-// ---------------------------------------------------------------------------
-
-const scopeFilterSchema = {
-	scope_id: z.string().optional().describe("Filter by a single sharing domain scope_id"),
-	include_scope_ids: z.array(z.string()).optional().describe("Sharing domain scope_ids to include"),
-	exclude_scope_ids: z.array(z.string()).optional().describe("Sharing domain scope_ids to exclude"),
-};
-
-const filterSchema = {
-	kind: z.string().optional().describe("Filter by memory kind"),
-	project: z.string().optional().describe("Filter by project scope (matches sessions.project)"),
-	...scopeFilterSchema,
-	visibility: z.array(z.string()).optional(),
-	include_visibility: z.array(z.string()).optional(),
-	exclude_visibility: z.array(z.string()).optional(),
-	include_workspace_ids: z.array(z.string()).optional(),
-	exclude_workspace_ids: z.array(z.string()).optional(),
-	include_workspace_kinds: z.array(z.string()).optional(),
-	exclude_workspace_kinds: z.array(z.string()).optional(),
-	include_actor_ids: z.array(z.string()).optional(),
-	exclude_actor_ids: z.array(z.string()).optional(),
-	include_trust_states: z.array(z.string()).optional(),
-	exclude_trust_states: z.array(z.string()).optional(),
-	ownership_scope: z.string().optional(),
-	personal_first: z.union([z.boolean(), z.string()]).optional(),
-	trust_bias: z.string().optional(),
-	widen_shared_when_weak: z.union([z.boolean(), z.string()]).optional(),
-	widen_shared_min_personal_results: z.number().int().optional(),
-	widen_shared_min_personal_score: z.number().optional(),
-};
-
-// ---------------------------------------------------------------------------
-// Helpers
-// ---------------------------------------------------------------------------
-
-const defaultProject = resolveDefaultProject();
-
-/** Wrap a JSON result into the MCP text content envelope. */
-function jsonContent(data: unknown) {
-	return { content: [{ type: "text" as const, text: JSON.stringify(data) }] };
-}
-
-/** Wrap an error message into the MCP text content envelope. */
-function errorContent(message: string) {
-	return { content: [{ type: "text" as const, text: JSON.stringify({ error: message }) }] };
-}
-
-// ---------------------------------------------------------------------------
-// Viewer autostart
-// ---------------------------------------------------------------------------
-
-/** Default viewer host/port for health checks. */
-const VIEWER_HOST = process.env.CODEMEM_VIEWER_HOST ?? "127.0.0.1";
-const VIEWER_PORT = process.env.CODEMEM_VIEWER_PORT ?? "38888";
-
-/** Resolve the `codemem` CLI binary path. Checks package-local paths first, then PATH. */
-function resolveCliPath(): string | null {
-	// When installed via npm/npx, the MCP server and CLI share node_modules.
-	// Walk up from this file to find the CLI's bin entry.
-	const candidates: string[] = [];
-	const selfDir = dirname(import.meta.dirname ?? ".");
-	// Monorepo dev: sibling package
-	candidates.push(join(selfDir, "..", "cli", "dist", "index.js"));
-	// npm install: node_modules/.bin/codemem
-	candidates.push(join(selfDir, "..", ".bin", "codemem"));
-	// npm install: deeper node_modules
-	candidates.push(join(selfDir, "..", "..", ".bin", "codemem"));
-	for (const p of candidates) {
-		if (existsSync(p)) return p;
-	}
-	// Fall back to PATH
-	return "codemem";
-}
-
-/** Check if the viewer is already running. */
-async function isViewerHealthy(): Promise<boolean> {
-	try {
-		const url = `http://${VIEWER_HOST}:${VIEWER_PORT}/api/health`;
-		const response = await fetch(url, { signal: AbortSignal.timeout(2_000) });
-		return response.ok;
-	} catch {
-		return false;
-	}
-}
-
-/**
- * Attempt to start the viewer server as a detached background process.
- * Best-effort: failures are logged to stderr but never block MCP startup.
- */
-async function ensureViewer(): Promise<void> {
-	if (process.env.CODEMEM_VIEWER === "0" || process.env.CODEMEM_VIEWER_AUTO === "0") return;
-
-	if (await isViewerHealthy()) return;
-
-	const cli = resolveCliPath();
-	if (!cli) return;
-
-	try {
-		// If the resolved path is a .js file, run it with node.
-		// Otherwise it's a bin script that can run directly.
-		const isJsFile = cli.endsWith(".js");
-		const cmd = isJsFile ? process.execPath : cli;
-		const args = isJsFile ? [cli, "serve", "start"] : ["serve", "start"];
-		// Pass non-default host/port so the spawned viewer matches what we health-check.
-		if (VIEWER_HOST !== "127.0.0.1") args.push("--host", VIEWER_HOST);
-		if (VIEWER_PORT !== "38888") args.push("--port", VIEWER_PORT);
-
-		const child = spawn(cmd, args, {
-			detached: true,
-			stdio: "ignore",
-			env: { ...process.env, CODEMEM_PLUGIN_IGNORE: "1" },
-		});
-		child.on("error", () => {}); // swallow — best effort
-		child.unref();
-
-		// Brief health check loop — don't block for long
-		for (let i = 0; i < 5; i++) {
-			await new Promise((r) => setTimeout(r, 1_000));
-			if (await isViewerHealthy()) return;
-		}
-	} catch {
-		// Best effort — MCP server continues regardless
-	}
-}
-
-// ---------------------------------------------------------------------------
-// Server bootstrap
-// ---------------------------------------------------------------------------
-
-async function main() {
-	const dbPath = resolveDbPath();
-	const store = new MemoryStore(dbPath);
-
-	// Auto-start viewer in background (belt and suspenders with any hook-based start)
-	ensureViewer().catch(() => {});
-
-	const server = new McpServer({ name: "codemem", version: VERSION });
-
-	// -------------------------------------------------------------------
-	// 1. memory_search
-	// -------------------------------------------------------------------
-	server.tool(
-		"memory_search",
-		"Search memories by text query. Returns full body text for each match.",
-		{
-			query: z.string().describe("Search query"),
-			limit: z.number().int().min(1).max(50).default(5).describe("Max results"),
-			...filterSchema,
-		},
-		async (args) => {
-			try {
-				const filters = buildFilters(args);
-				const items = store.search(args.query, args.limit, filters);
-				return jsonContent({
-					items: items.map((m: MemoryResult) => ({
-						id: m.id,
-						title: m.title,
-						kind: m.kind,
-						body: m.body_text,
-						confidence: m.confidence,
-						score: m.score,
-						session_id: m.session_id,
-						metadata: m.metadata,
-					})),
-				});
-			} catch (err) {
-				return errorContent(err instanceof Error ? err.message : String(err));
-			}
-		},
-	);
-
-	// -------------------------------------------------------------------
-	// 2. memory_search_index
-	// -------------------------------------------------------------------
-	server.tool(
-		"memory_search_index",
-		"Search memories by text query. Returns compact index entries (no body) for browsing.",
-		{
-			query: z.string().describe("Search query"),
-			limit: z.number().int().min(1).max(50).default(8).describe("Max results"),
-			...filterSchema,
-		},
-		async (args) => {
-			try {
-				const filters = buildFilters(args);
-				const items = store.search(args.query, args.limit, filters);
-				return jsonContent({
-					items: items.map((m: MemoryResult) => ({
-						id: m.id,
-						kind: m.kind,
-						title: m.title,
-						score: m.score,
-						created_at: m.created_at,
-						session_id: m.session_id,
-						metadata: m.metadata,
-					})),
-				});
-			} catch (err) {
-				return errorContent(err instanceof Error ? err.message : String(err));
-			}
-		},
-	);
-
-	// -------------------------------------------------------------------
-	// 3. memory_timeline
-	// -------------------------------------------------------------------
-	server.tool(
-		"memory_timeline",
-		"Get a chronological window of memories around an anchor (by ID or query).",
-		{
-			query: z.string().optional().describe("Search query to find anchor"),
-			memory_id: z.number().int().optional().describe("Anchor memory ID"),
-			depth_before: z.number().int().min(0).default(3).describe("Items before anchor"),
-			depth_after: z.number().int().min(0).default(3).describe("Items after anchor"),
-			...filterSchema,
-		},
-		async (args) => {
-			try {
-				const filters = buildFilters(args);
-				const items = store.timeline(
-					args.query ?? null,
-					args.memory_id ?? null,
-					args.depth_before,
-					args.depth_after,
-					filters,
-				);
-				return jsonContent({ items });
-			} catch (err) {
-				return errorContent(err instanceof Error ? err.message : String(err));
-			}
-		},
-	);
-
-	// -------------------------------------------------------------------
-	// 4. memory_explain
-	// -------------------------------------------------------------------
-	server.tool(
-		"memory_explain",
-		"Explain search results with detailed scoring breakdown.",
-		{
-			query: z.string().optional().describe("Search query"),
-			ids: z.array(z.number().int()).max(200).optional().describe("Specific memory IDs to explain"),
-			limit: z.number().int().min(1).max(50).default(10).describe("Max results"),
-			include_pack_context: z.boolean().default(false).describe("Include formatted pack context"),
-			...filterSchema,
-		},
-		async (args) => {
-			try {
-				const filters = buildFilters(args);
-				const result = store.explain(args.query ?? null, args.ids ?? null, args.limit, filters, {
-					includePackContext: args.include_pack_context,
-				});
-				return jsonContent(result);
-			} catch (err) {
-				return errorContent(err instanceof Error ? err.message : String(err));
-			}
-		},
-	);
-
-	// -------------------------------------------------------------------
-	// 5. memory_expand
-	// -------------------------------------------------------------------
-	server.tool(
-		"memory_expand",
-		"Fetch memories by ID with surrounding timeline context.",
-		{
-			ids: z
-				.array(z.union([z.number(), z.string()]))
-				.max(200)
-				.describe("Memory IDs to expand"),
-			depth_before: z.number().int().min(0).default(3).describe("Timeline items before"),
-			depth_after: z.number().int().min(0).default(3).describe("Timeline items after"),
-			include_observations: z.boolean().default(false).describe("Include full observation details"),
-			...filterSchema,
-		},
-		async (args) => {
-			try {
-				// Python: project if project is not None else default_project
-				// Explicit "" clears project scoping; only fall back to default when undefined.
-				const filterDefaultProject =
-					args.project !== undefined && !args.project.trim() ? null : defaultProject;
-				const filters = buildFilters(args, filterDefaultProject);
-				const resolvedProject = filters?.project ?? null;
-				const { ordered: orderedIds, invalid: invalidIds } = dedupeOrderedIds(args.ids);
-				const errors: Array<Record<string, unknown>> = [];
-
-				if (invalidIds.length > 0) {
-					errors.push({
-						code: "INVALID_ARGUMENT",
-						field: "ids",
-						message: "some ids are not valid integers",
-						ids: invalidIds,
-					});
-				}
-
-				const missingNotFound: number[] = [];
-				const missingProjectMismatch: number[] = [];
-				const missingFilterMismatch: number[] = [];
-				const anchors: MemoryItemResponse[] = [];
-				const timelineItems: MemoryItemResponse[] = [];
-				const timelineSeen = new Set<number>();
-				const sessionProjects = new Map<number, string | null>();
-
-				for (const memoryId of orderedIds) {
-					const item = store.get(memoryId);
-					if (!item?.active) {
-						missingNotFound.push(memoryId);
-						continue;
-					}
-
-					const sessionId = item.session_id;
-					if (resolvedProject && sessionId > 0) {
-						if (!sessionProjects.has(sessionId)) {
-							const row = store.db
-								.prepare("SELECT project FROM sessions WHERE id = ? LIMIT 1")
-								.get(sessionId) as { project: string | null } | undefined;
-							sessionProjects.set(sessionId, typeof row?.project === "string" ? row.project : null);
-						}
-						if (!projectMatchesFilter(resolvedProject, sessionProjects.get(sessionId) ?? null)) {
-							missingProjectMismatch.push(memoryId);
-							continue;
-						}
-					} else if (resolvedProject && sessionId <= 0) {
-						missingProjectMismatch.push(memoryId);
-						continue;
-					}
-
-					const expanded = store.timeline(
-						null,
-						memoryId,
-						args.depth_before,
-						args.depth_after,
-						filters,
-					);
-					const anchor = expanded.find((expandedItem) => expandedItem.id === memoryId);
-					if (!anchor) {
-						missingFilterMismatch.push(memoryId);
-						continue;
-					}
-
-					anchors.push(anchor);
-					for (const expandedItem of expanded) {
-						const expandedId = expandedItem.id;
-						if (expandedId <= 0 || timelineSeen.has(expandedId)) continue;
-						timelineSeen.add(expandedId);
-						timelineItems.push(expandedItem);
-					}
-				}
-
-				if (missingNotFound.length > 0) {
-					errors.push({
-						code: "NOT_FOUND",
-						field: "ids",
-						message: "some requested ids were not found",
-						ids: missingNotFound,
-					});
-				}
-				if (missingProjectMismatch.length > 0) {
-					errors.push({
-						code: "PROJECT_MISMATCH",
-						field: "project",
-						message: "some requested ids are outside the requested project scope",
-						ids: missingProjectMismatch,
-					});
-				}
-				if (missingFilterMismatch.length > 0) {
-					errors.push({
-						code: "FILTER_MISMATCH",
-						field: "filters",
-						message: "some requested ids are outside the requested filters",
-						ids: missingFilterMismatch,
-					});
-				}
-
-				// Collect full observations if requested
-				let observations: MemoryItemResponse[] = [];
-				if (args.include_observations) {
-					const observationSeen = new Set<number>();
-					const observationIds: number[] = [];
-					for (const item of [...anchors, ...timelineItems]) {
-						if (item.id > 0 && !observationSeen.has(item.id)) {
-							observationSeen.add(item.id);
-							observationIds.push(item.id);
-						}
-					}
-					observations = getManyForMcp(store, observationIds, filters);
-				}
-
-				return jsonContent({
-					anchors,
-					timeline: timelineItems,
-					observations,
-					missing_ids: orderedIds.filter(
-						(memoryId: number) =>
-							missingNotFound.includes(memoryId) ||
-							missingProjectMismatch.includes(memoryId) ||
-							missingFilterMismatch.includes(memoryId),
-					),
-					errors,
-					metadata: {
-						project: resolvedProject,
-						requested_ids_count: orderedIds.length,
-						returned_anchor_count: anchors.length,
-						timeline_count: timelineItems.length,
-						include_observations: args.include_observations,
-					},
-				});
-			} catch (err) {
-				return errorContent(err instanceof Error ? err.message : String(err));
-			}
-		},
-	);
-
-	// -------------------------------------------------------------------
-	// 6. memory_get
-	// -------------------------------------------------------------------
-	server.tool(
-		"memory_get",
-		"Fetch a single memory item by ID.",
-		{
-			memory_id: z.number().int().describe("Memory ID"),
-			...filterSchema,
-		},
-		async (args) => {
-			try {
-				const item = getMemoryForMcp(store, args.memory_id, buildFilters(args, null));
-				if (!item) return errorContent("not_found");
-				return jsonContent(item);
-			} catch (err) {
-				return errorContent(err instanceof Error ? err.message : String(err));
-			}
-		},
-	);
-
-	// -------------------------------------------------------------------
-	// 7. memory_get_observations
-	// -------------------------------------------------------------------
-	server.tool(
-		"memory_get_observations",
-		"Fetch multiple memory items by their IDs.",
-		{
-			ids: z.array(z.number().int()).max(200).describe("Memory IDs to fetch"),
-			...filterSchema,
-		},
-		async (args) => {
-			try {
-				const items = getManyForMcp(store, args.ids, buildFilters(args, null));
-				return jsonContent({ items });
-			} catch (err) {
-				return errorContent(err instanceof Error ? err.message : String(err));
-			}
-		},
-	);
-
-	// -------------------------------------------------------------------
-	// 8. memory_recent
-	// -------------------------------------------------------------------
-	server.tool(
-		"memory_recent",
-		"Return recent memories, newest first.",
-		{
-			limit: z.number().int().min(1).max(100).default(8).describe("Max results"),
-			...filterSchema,
-		},
-		async (args) => {
-			try {
-				const filters = buildFilters(args);
-				const items = store.recent(args.limit, filters);
-				return jsonContent({ items });
-			} catch (err) {
-				return errorContent(err instanceof Error ? err.message : String(err));
-			}
-		},
-	);
-
-	// -------------------------------------------------------------------
-	// 9. memory_pack
-	// -------------------------------------------------------------------
-	server.tool(
-		"memory_pack",
-		"Build a formatted memory pack from search results — quick one-shot context block.",
-		{
-			context: z.string().describe("Context description to search for"),
-			limit: z.number().int().min(1).max(50).optional().describe("Max items to include"),
-			compact: z
-				.boolean()
-				.optional()
-				.describe(
-					"When true, render a scannable index of all items with full detail only for the top N (default 3). Saves tokens when broad overview matters more than per-item detail.",
-				),
-			compact_detail_count: z
-				.number()
-				.int()
-				.min(0)
-				.max(50)
-				.optional()
-				.describe("Number of items to show in full detail in compact mode (default 3)"),
-			compression_mode: z
-				.enum(["off", "compact", "ids"])
-				.optional()
-				.describe(
-					"Near-related compression mode: off disables it, compact applies only to compact rendering, ids applies in all modes. Defaults to CODEMEM_PACK_COMPRESSION or compact.",
-				),
-			...filterSchema,
-		},
-		async (args) => {
-			try {
-				const filters = buildFilters(args);
-				const renderOptions =
-					args.compact || args.compact_detail_count != null || args.compression_mode != null
-						? {
-								compact: args.compact ?? (args.compact_detail_count != null ? true : undefined),
-								compactDetailCount: args.compact_detail_count,
-								compressionMode: args.compression_mode,
-							}
-						: undefined;
-				const result = await store.buildMemoryPackAsync(
-					args.context,
-					args.limit ?? undefined,
-					null,
-					filters,
-					renderOptions,
-				);
-				return jsonContent(result);
-			} catch (err) {
-				return errorContent(err instanceof Error ? err.message : String(err));
-			}
-		},
-	);
-
-	// -------------------------------------------------------------------
-	// 10. memory_remember
-	// -------------------------------------------------------------------
-	server.tool(
-		"memory_remember",
-		"Create a new memory. Use for milestones, decisions, and notable facts.",
-		{
-			kind: z
-				.enum(["discovery", "change", "feature", "bugfix", "refactor", "decision", "exploration"])
-				.describe("Memory kind"),
-			title: z.string().describe("Short title"),
-			body: z.string().describe("Body text (high-signal content)"),
-			confidence: z.number().min(0).max(1).default(0.5).describe("Confidence 0-1"),
-			project: z.string().optional().describe("Project identifier"),
-		},
-		async (args) => {
-			try {
-				const result = rememberMemoryForMcp(store, args, {
-					envProject: process.env.CODEMEM_PROJECT ?? null,
-				});
-
-				try {
-					await storeVectors(store.db, result.memId, result.title, result.body);
-				} catch {
-					// Non-fatal — memory writes should succeed even if embeddings are unavailable
-				}
-
-				return jsonContent({ id: result.memId });
-			} catch (err) {
-				return errorContent(err instanceof Error ? err.message : String(err));
-			}
-		},
-	);
-
-	// -------------------------------------------------------------------
-	// 11. memory_forget
-	// -------------------------------------------------------------------
-	server.tool(
-		"memory_forget",
-		"Soft-delete a memory item. Use for incorrect or sensitive data.",
-		{
-			memory_id: z.number().int().describe("Memory ID to forget"),
-			...filterSchema,
-		},
-		async (args) => {
-			try {
-				if (!forgetMemoryForMcp(store, args.memory_id, buildFilters(args, null))) {
-					return errorContent("not_found");
-				}
-				return jsonContent({ status: "ok" });
-			} catch (err) {
-				return errorContent(err instanceof Error ? err.message : String(err));
-			}
-		},
-	);
-
-	// -------------------------------------------------------------------
-	// 12. memory_schema
-	// -------------------------------------------------------------------
-	server.tool(
-		"memory_schema",
-		"Return the memory schema — kinds, fields, and available filters.",
-		{},
-		async () => {
-			return jsonContent({
-				kinds: Object.keys(MEMORY_KINDS),
-				kind_descriptions: MEMORY_KINDS,
-				fields: {
-					title: "short text",
-					body: "long text",
-					subtitle: "short text",
-					facts: "list<string>",
-					narrative: "long text",
-					concepts: "list<string>",
-					files_read: "list<string>",
-					files_modified: "list<string>",
-					prompt_number: "int",
-				},
-				filters: [
-					"kind",
-					"session_id",
-					"since",
-					"project",
-					"scope_id",
-					"include_scope_ids",
-					"exclude_scope_ids",
-					"include_actor_ids",
-					"exclude_actor_ids",
-					"include_visibility",
-					"exclude_visibility",
-					"include_workspace_ids",
-					"exclude_workspace_ids",
-					"include_workspace_kinds",
-					"exclude_workspace_kinds",
-				],
-			});
-		},
-	);
-
-	// -------------------------------------------------------------------
-	// 13. memory_learn
-	// -------------------------------------------------------------------
-	server.tool(
-		"memory_learn",
-		"Learn how to use codemem memory tools. Call this first if unfamiliar.",
-		{},
-		async () => {
-			return jsonContent({
-				intro: "Use this tool when you're new to codemem or unsure when to recall/persist.",
-				client_hint: "If you are unfamiliar with codemem, call memory.learn first.",
-				recall: {
-					when: [
-						"Start of a task or when the user references prior work.",
-						"When you need background context, decisions, or recent changes.",
-					],
-					how: [
-						"Use memory.search_index to get compact candidates.",
-						"Use memory.timeline to expand around a promising memory.",
-						"Use memory.get_observations for full details only when needed.",
-						"Use memory.pack for quick one-shot context blocks.",
-						"Use the project filter unless the user requests cross-project context.",
-					],
-					examples: [
-						'memory.search_index("billing cache bug", limit=5)',
-						"memory.timeline(memory_id=123)",
-						"memory.get_observations([123, 456])",
-					],
-				},
-				persistence: {
-					when: [
-						"Milestones (task done, key decision, new facts learned).",
-						"Notable regressions or follow-ups that should be remembered.",
-					],
-					how: [
-						"Use memory.remember with kind decision/discovery/change/exploration.",
-						"Keep titles short and bodies high-signal.",
-						"ALWAYS pass the project parameter if known.",
-					],
-					examples: [
-						'memory.remember(kind="decision", title="Switch to async cache", body="...why...", project="my-service")',
-						'memory.remember(kind="change", title="Fixed retry loop", body="...impact...", project="my-service")',
-					],
-				},
-				forget: {
-					when: [
-						"Accidental or sensitive data stored in memory items.",
-						"Obsolete or incorrect items that should no longer surface.",
-					],
-					how: [
-						"Call memory.forget(id) to mark the item inactive.",
-						"Prefer forgetting over overwriting to preserve auditability.",
-					],
-					examples: ["memory.forget(123)"],
-				},
-				prompt_hint:
-					"At task start: call memory.search_index; during work: memory.timeline + memory.get_observations; at milestones: memory.remember.",
-				recommended_system_prompt: [
-					"Trigger policy (1-liner): If the user references prior work or starts a task,",
-					"immediately call memory.search_index; then use memory.timeline + memory.get_observations;",
-					"at milestones, call memory.remember; use memory.forget for incorrect/sensitive items.",
-					"",
-					"System prompt:",
-					"You have access to codemem MCP tools. If unfamiliar, call memory.learn first.",
-					"",
-					"Recall:",
-					"- Start of any task: call memory.search_index with a concise task query.",
-					'- If prior work is referenced ("as before", "last time", "we already did…", "regression"),',
-					"  call memory.search_index or memory.timeline.",
-					"- Use memory.get_observations only after filtering IDs.",
-					"- Prefer project-scoped queries unless the user asks for cross-project.",
-					"",
-					"Persistence:",
-					"- On milestones (task done, key decision, new facts learned), call memory.remember.",
-					"- Use kind=decision for tradeoffs, kind=change for outcomes, kind=discovery/exploration for useful findings.",
-					"- Keep titles short and bodies high-signal.",
-					"- ALWAYS pass the project parameter if known.",
-					"",
-					"Safety:",
-					"- Use memory.forget(id) for incorrect or sensitive items.",
-					"",
-					"Examples:",
-					'- memory.search_index("billing cache bug")',
-					"- memory.timeline(memory_id=123)",
-					"- memory.get_observations([123, 456])",
-					'- memory.remember(kind="decision", title="Use async cache", body="Chose async cache to avoid lock contention in X.", project="my-service")',
-					'- memory.remember(kind="change", title="Fixed retry loop", body="Root cause was Y; added guard in Z.", project="my-service")',
-					"- memory.forget(123)",
-				].join("\n"),
-			});
-		},
-	);
-
-	// -------------------------------------------------------------------
-	// Connect and handle shutdown
-	// -------------------------------------------------------------------
-	const transport = new StdioServerTransport();
-
-	const shutdown = () => {
-		try {
-			store.close();
-		} catch {
-			// Best effort — process is exiting
-		}
-		process.exit(0);
-	};
-
-	process.on("SIGINT", shutdown);
-	process.on("SIGTERM", shutdown);
-
-	await server.connect(transport);
-}
-
-main().catch((err) => {
-	console.error("codemem MCP server failed to start:", err);
-	process.exit(1);
-});
+export { createCodememMcpServer } from "./server.js";
+export type { CodememMcpServerOptions } from "./tool-context.js";
