+++ packages/mcp-server/src/memory-kinds.ts
@@ -1,0 +1,9 @@
+export const MEMORY_KINDS: Record<string, string> = {
+	discovery: "Something learned about the codebase, architecture, or tools",
+	change: "A code change that was made",
+	feature: "A new feature that was implemented",
+	bugfix: "A bug that was found and fixed",
+	refactor: "Code that was refactored or restructured",
+	decision: "A design or architecture decision",
+	exploration: "An experiment or investigation (may not have shipped)",
+};
